--- src/components/Main/components/Content/components/FullWideContext/components/ArticleDetail/StudentDetail/StudentDetail.module.scss
@@ -1,5 +1,0 @@
-@import './../../../../../../../../../variable';
-
-.wrapper{
-  padding:20px;
-}
