+++ src/components/Main/components/Content/components/FullWidthLayout/components/StudentDetail/StudentDetail.module.scss
@@ -1,0 +1,5 @@
+@import './../../../../content-variable';
+
+.wrapper{
+  padding:20px;
+}
