+++ src/components/Main/components/Content/components/FullWidthLayout/components/StudentDetail/index.js
